Update agenda.jsx
Agenda Page now handles deletion of events from the database. Press and hold the event on the specified date to remove.

Issues: Agenda is not able to re-render changes immediately. Upon deletion (long press), we need to move back to the calendar page and re-enter into the agenda page to see that the event has been deleted from the database